# backend/openapi2-hermes-v2.yaml
swagger: '2.0'
info:
  description: "The specification of a Hermes backend API for frontend and third party software"
  version: "1.8.2"
  title: "SocialEdge Hermes API"
  contact:
    email: "socialedge.team@gmail.com"
  license:
    name: "GNU General Public License, v3"
    url: "https://www.gnu.org/licenses/gpl-3.0.txt"
schemes:
  - "http"
produces:
  - "application/vnd.socialedge.hermes.v2+json"
...
paths:
...
  /stations:
    get:
      tags:
        - stations
      operationId: "listStations"
      summary: Lists all Stations from the system that the user has access to
      parameters:
        - name: "size"
          in: "query"
          description: "Limits an amount of entities per page"
          required: false
          type: "integer"
        - name: "page"
          in: "query"
          description: "Number of list page to display"
          required: false
          type: "integer"
        - name: "sort"
          in: "query"
          description: "Defines a sort params for the query e.g ?sort=name,ASC"
          required: false
          type: "string"
      responses:
        200:
          description: A list of Stations
          headers:
            "X-Page-Size":
              description: Amount of elements per page if page was set by query parameter
              type: integer
            "X-Page-Number":
              description: A page number if page was set by query parameter
              type: integer
            "X-Page-Total":
              description: A total amount of pages required to page all resources with given X-Page-Size
              type: integer
            "X-Resource-Total-Records":
              description: A total amount of records residing on this endpoint
              type: integer
          schema:
            type: array
            minItems: 0
            items:
              $ref: '#/definitions/StationDTO'
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
    post:
      tags:
        - stations
      operationId: "createStation"
      summary: Creates new Station
      parameters:
        - name: body
          in: body
          description: Station to add to the store
          required: true
          schema:
            $ref: '#/definitions/StationDTO'
      responses:
        201:
          description: A Station created
          schema:
            $ref: '#/definitions/StationDTO'
        400:
          description: Problem with the request, such as a missing, invalid or type mismatched parameter
          schema:
            $ref: '#/definitions/ErrorDTO'
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
  /stations/{id}:
    get:
      tags:
        - stations
      operationId: "getStation"
      summary: Returns a Stations based on a single ID
      parameters:
        - name: id
          in: path
          description: ID of a Station to fetch
          required: true
          type: string
      responses:
        200:
          description: Line fetched
          schema:
            $ref: '#/definitions/StationDTO'
        404:
          description: Station not found
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
    put:
      tags:
        - stations
      operationId: "replaceStation"
      summary: Updates a Station
      parameters:
        - name: id
          in: path
          description: ID of a Station to update
          required: true
          type: string
        - name: body
          in: body
          description: Full Station body to update existing one
          required: true
          schema:
            $ref: '#/definitions/StationDTO'
      responses:
        200:
          description: Station created
          schema:
            $ref: '#/definitions/StationDTO'
        404:
          description: Station not found
        400:
          description: Problem with the request, such as a missing, invalid or type mismatched parameter
          schema:
            $ref: '#/definitions/ErrorDTO'
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
    delete:
      tags:
        - stations
      operationId: "deleteStation"
      summary: Deletes a single Station based on the ID supplied
      parameters:
        - name: id
          in: path
          description: ID of a Station to delete
          required: true
          type: string
      responses:
        204:
          description: Station deleted
        404:
          description: Station not found
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
  /lines:
    get:
      tags:
        - lines
      operationId: "listLines"
      summary: Lists all lines from the system that the user has access to
      parameters:
        - name: "size"
          in: "query"
          description: "Limits an amount of entities per page"
          required: false
          type: "integer"
        - name: "page"
          in: "query"
          description: "Number of list page to display"
          required: false
          type: "integer"
        - name: "sort"
          in: "query"
          description: "Defines a sort params for the query e.g ?sort=name,ASC"
          required: false
          type: "string"
      responses:
        200:
          description: A list of Lines
          headers:
            "X-Page-Size":
              description: Amount of elements per page if page was set by query parameter
              type: integer
            "X-Page-Number":
              description: A page number if page was set by query parameter
              type: integer
            "X-Page-Total":
              description: A total amount of pages required to page all resources with given X-Page-Size
              type: integer
            "X-Resource-Total-Records":
              description: A total amount of records residing on this endpoint
              type: integer
          schema:
            type: array
            minItems: 0
            items:
              $ref: '#/definitions/LineDTO'
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
    post:
      tags:
        - lines
      operationId: "createLine"
      summary: Creates new Line
      parameters:
        - name: body
          in: body
          description: Line to add to the store
          required: true
          schema:
            $ref: '#/definitions/LineDTO'
      responses:
        201:
          description: A Line created
          schema:
            $ref: '#/definitions/LineDTO'
        400:
          description: Problem with the request, such as a missing, invalid or type mismatched parameter
          schema:
            $ref: '#/definitions/ErrorDTO'
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
  /lines/{id}:
    get:
      tags:
        - lines
      operationId: "getLine"
      summary: Returns a Line based on a single ID
      parameters:
        - name: id
          in: path
          description: ID of a Line to fetch
          required: true
          type: string
      responses:
        200:
          description: Line fetched
          schema:
            $ref: '#/definitions/LineDTO'
        404:
          description: Line not found
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
    put:
      tags:
        - lines
      operationId: "replaceLine"
      summary: Updates a Line
      parameters:
        - name: id
          in: path
          description: ID of a Line to update
          required: true
          type: string
        - name: body
          in: body
          description: Full Line body to update existing one
          required: true
          schema:
            $ref: '#/definitions/LineDTO'
      responses:
        200:
          description: Line created
          schema:
            $ref: '#/definitions/LineDTO'
        404:
          description: Line not found
        400:
          description: Problem with the request, such as a missing, invalid or type mismatched parameter
          schema:
            $ref: '#/definitions/ErrorDTO'
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
    delete:
      tags:
        - lines
      operationId: "deleteLine"
      summary: Deletes a single Line based on the ID supplied
      parameters:
        - name: id
          in: path
          description: ID of an Line to delete
          required: true
          type: string
      responses:
        204:
          description: Line deleted
        404:
          description: Line not found
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
...
  /tasks/shedules:
    post:
      tags:
        - tasks
      operationId: "generateSchedule"
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/ScheduleGenReqDTO'
      responses:
        200:
          description: Information about task queried
          schema:
            $ref: '#/definitions/ScheduleGenTaskDTO'
        400:
          description: Problem with the request, such as a missing, invalid or type mismatched parameter
          schema:
            $ref: '#/definitions/ErrorDTO'
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
  /tasks/schedules/{id}:
    get:
      tags:
        - tasks
      operationId: "getScheduleGenerationStatus"
      parameters:
        - name: id
          in: path
          description: ID of a Schedule Generation Task
          required: true
          type: string
      responses:
        200:
          description: Information about task queried
          schema:
            $ref: '#/definitions/ScheduleGenTaskDTO'
        400:
          description: Problem with the request, such as a missing, invalid or type mismatched parameter
          schema:
            $ref: '#/definitions/ErrorDTO'
        500:
          description: Internal server error, please contact support
          schema:
            $ref: '#/definitions/ErrorDTO'
...
definitions:
  LocationDTO:
    type: object
    required:
    - latitude
    - longitude
    properties:
      latitude:
        type: number
        format: double
        minimum: 90
        maximum: 90
      longitude:
        type: number
        format: double
        minimum: 180
        maximum: 180
...
  StationDTO:
    type: object
    required:
    - name
    - vehicleType
    - location
    - dwell
    description: >-
      A stop is a location where vehicles stop to pick up or drop off
      passengers.
    properties:
      id:
        type: string
        readOnly: true
        minLength: 36
        maxLength: 36
      name:
        type: string
        minLength: 1
      description:
        type: string
      vehicleType:
        type: array
        minItems: 1
        items:
          type: string
      location:
        $ref: '#/definitions/LocationDTO'
      dwell:
        type: string
        description: Dwell time in ISO-8601 duration format (PnDTnHnMn.nS)
        minLength: 2
  LineDTO:
    type: object
    required:
    - name
    - vehicleType
    - agencyId
    - inboundRoute
    description: >-
      Lines represents a group of Routes that are displayed to riders as a
      single service.
    properties:
      id:
        type: string
        readOnly: true
        minLength: 36
        maxLength: 36
      agencyId:
        type: string
        minLength: 36
        maxLength: 36
      name:
        type: string
        minLength: 1
      description:
        type: string
      vehicleType:
        type: string
        minLength: 1
      url:
        type: string
      inboundRoute:
        $ref: '#/definitions/RouteDTO'
      outboundRoute:
        $ref: '#/definitions/RouteDTO'
  RouteDTO:
    type: array
    minItems: 1
    uniqueItems: true
    description: >-
      Transit Routes define waypoints for a journey taken by a vehicle along a
      transit line.
    items:
      $ref: '#/definitions/SegmentDTO'
  SegmentDTO:
    type: object
    description: 'Represents an edge of the journey graph, defined by Route'
    properties:
      begin:
        $ref: '#/definitions/SegmentVertexDTO'
      end:
        $ref: '#/definitions/SegmentVertexDTO'
      length:
        type: number
        format: double
      waypoints:
        type: array
        description: Contains all waypoints required to travel from Segment start to end
        items:
          $ref: '#/definitions/LocationDTO'
  SegmentVertexDTO:
    type: object
    description: 'Represents a vertex (stop) of a Segment'
    required:
      - stationId
    properties:
      stationId:
        type: string
      name:
        type: string
      location:
        $ref: '#/definitions/LocationDTO'
...
  AvailabilityDTO:
    type: object
    required:
      - dayOfWeek
      - startDate
    description: >-
      Defines a range of dates between which the {@link Schedule} is available
      and the days of the week when it is available (such as Monday through
      Friday). It also may define specific days when a trip is not available,
      such as holidays.
    properties:
      dayOfWeek:
        type: array
        minItems: 1
        items:
          type: string
      startDate:
        type: string
        format: date
      endDate:
        type: string
        format: date
      exceptionDays:
        type: array
        items:
          type: string
          format: date
...
  ScheduleGenReqDTO:
    type: object
    description: Specification of a schedule to be generated
    properties:
      description:
        type: string
      lineId:
        type: string
        minLength: 36
        maxLength: 36
      availability:
        $ref: '#/definitions/AvailabilityDTO'
      startTimeInbound:
        type: string
        description: Local time in ISO-8601 format (HH:mm - HH:mm:ss.SSSSSSSSS)
        minLength: 5
      endTimeInbound:
        type: string
        description: Local time in ISO-8601 format (HH:mm - HH:mm:ss.SSSSSSSSS)
        minLength: 5
      startTimeOutbound:
        type: string
        description: Local time in ISO-8601 format (HH:mm - HH:mm:ss.SSSSSSSSS)
        minLength: 5
      endTimeOutbound:
        type: string
        description: Local time in ISO-8601 format (HH:mm - HH:mm:ss.SSSSSSSSS)
        minLength: 5
      headway:
        type: string
        description: Headway in ISO-8601 duration format (PnDTnHnMn.nS)
        minLength: 2
      averageSpeed:
        type: string
        minLength: 1
      minLayover:
        type: string
        description: Headway in ISO-8601 duration format (PnDTnHnMn.nS)
        minLength: 2
  TaskDTO:
    type: object
    description: Carries info about queried task
    required:
      - id
      - status
      - started
    properties:
      id:
        type: string
        minLength: 36
        maxLength: 36
      status:
        type: string
        description: QUERIED/IN_PROCESS/DONE
      started:
        type: string
        format: date-time
      finished:
        type: string
        format: date-time
  ScheduleGenTaskDTO:
    type: object
    allOf:
      - $ref: "#/definitions/TaskDTO"
      - type: object
        properties:
          scheduleId:
            type: string
            minLength: 36
            maxLength: 36
  ErrorDTO:
    type: object
    description: Describes a failure occured
    properties:
      timestamp:
        type: string
        format: date
      status:
        type: integer
      error:
        type: string
      exception:
        type: string
        description: A java exception class name
      message:
        type: string
      path:
        type: string